Segregated tests by Sagemaker entity.  Model arn check by regex..
